Camera Trajectory Working (only one trajectory can be defined in ogitor for the moment)

+++ OUAN/OUAN/Src/Graphics/CameraManager/CameraController.cpp
@@ -25,7 +25,7 @@ Camera * CameraController::getCamera()
 	return mCamera;
 }
 
-void CameraController::update(long elapsedTime)
+void CameraController::update(double elapsedTime)
 {
 
 }
